updated imports, beam parameters and particle generation
