+++ fma_ions/__init__.py
@@ -1,13 +1,13 @@
 from .fma_ions import FMA
 
 from .sequences import SPS_sequence_maker
-from .sequences import BeamParameters_SPS, BeamParameters_SPS_Binomial_2016, BeamParameters_SPS_Binomial_2016_before_RF_Spill, BeamParameters_SPS_Oxygen, BeamParameters_SPS_Proton
-
 from .sequences import PS_sequence_maker, BeamParameters_PS
 from .sequences import LEIR_sequence_maker, BeamParameters_LEIR
 
+from .beam_parameters import BeamParameters_SPS, BeamParameters_SPS_Binomial_2016, BeamParameters_SPS_Binomial_2016_before_RF_capture, BeamParameters_SPS_Oxygen, BeamParameters_SPS_Proton
+
+
 from .resonance_lines import resonance_lines
-from .dynamic_aperture.dynamic_aperture import Dynamic_Aperture
 from .tune_ripple import Tune_Ripple_SPS
 from .sps_flat_bottom_tracking import SPS_Flat_Bottom_Tracker
 from .submitter import Submitter
