+++ fma_ions/beam_parameters/__init__.py
@@ -1,0 +1,1 @@
+from .sps_beam_parameters import BeamParameters_SPS, BeamParameters_SPS_Binomial_2016, BeamParameters_SPS_Binomial_2016_before_RF_capture, BeamParameters_SPS_Oxygen, BeamParameters_SPS_Proton
