+++ tests/flat_bottom_tracking/test_sps_tracking_class.py
@@ -5,7 +5,4 @@ import fma_ions
 
 # Test default tracking with space charge on GPU context - then test plotting
 sps = fma_ions.SPS_Flat_Bottom_Tracker(num_turns=100)
-tbt = sps.track_SPS(which_context='gpu', add_aperture=False)
-
-# Test loading the data correctly again
-tbt2 = sps.load_tbt_data()
+tbt = sps.track_SPS(which_context='cpu', add_aperture=False)
